requirements.txt
streamlit
google-generativeai
python-dotenv
ipython
xlsxwriter
openpyxl
google-cloud-storage
numpy
gcsfs